lazy load auth in login handlers
