+++ integrations/cypress/cypress/support/commands/authorization_code_with_pkce/index.ts
@@ -1,35 +1,31 @@
 import { Slice } from '@effection/atom';
 import { TestState } from '../../types';
 import { assert } from 'assert-ts';
+import { makeCypressLogger } from '../../utils/cypress-logger';
 
 export interface MakeLoginOptions {
   atom: Slice<TestState>;
 }
 
+const log = makeCypressLogger('simulacrum-login-pkce');
+
 export const makeLoginWithPKCE = ({ atom }: MakeLoginOptions) => () => {
   return new Cypress.Promise((resolve, reject) => {
-    let person = atom.slice(Cypress.spec.name, 'person').get();
+    import('./auth').then(m => m.auth0Client).then((auth0Client) => {
+      let person = atom.slice(Cypress.spec.name, 'person').get();
 
-    assert(!!person && typeof person.email !== 'undefined', `no scenario in login`);
+      assert(!!person && typeof person.email !== 'undefined', `no scenario in login`);
 
-    auth0Client.getTokenSilently({ ignoreCache: true, currentUser: person.email, test: Cypress.currentTest.title })
-               .then(token => {
-                  Cypress.log({
-                    name: 'simulacrum-login',
-                    displayName: 'simulacrum-login',
-                    message: `successfully logged in with token ${JSON.stringify(token)}`
-                  });
+      auth0Client.getTokenSilently({ ignoreCache: true, currentUser: person.email, test: Cypress.currentTest.title })
+      .then(token => {
+        log(`successfully logged in with token ${JSON.stringify(token)}`);
 
-                  resolve(token);
-               })
-               .catch((e) => {
-                Cypress.log({
-                  name: 'simulacrum-login',
-                  displayName: 'simulacrum-login',
-                  message: `login failed ${e.message}`
-                });
+        resolve(token);
+      }).catch((e) => {
+        console.error(e);
 
-                 reject(e);
-               });
+        reject(e);
+      });
+    });
   });
 };
