+++ integrations/cypress/cypress/support/commands/index.ts
@@ -5,11 +5,11 @@ import { makeCreateSimulation } from './create-simulation';
 import { CreateSimulation, Person, TestState, Token } from '../types';
 import { makeGetClientFromSpec } from '../utils/spec';
 import { makeGiven } from './given';
-import { makeLogin } from './authorization_code';
+import { makeLogin } from './login';
 import { makeLogout } from './logout';
 import { Auth0Result } from 'auth0-js';
 import './get-user-tokens';
-import './get-user-info';
+import './authorization_code/get-user-info';
 
 declare global {
   namespace Cypress {
@@ -37,7 +37,8 @@ Cypress.Commands.add('createSimulation', makeCreateSimulation({ atom, getClientF
 
 Cypress.Commands.add('given', makeGiven({ atom, getClientFromSpec }));
 
-Cypress.Commands.add('login', makeLogin({ atom }));
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+Cypress.Commands.add('login', makeLogin({ atom }) as any);
 
 Cypress.Commands.add('logout', makeLogout({ atom, getClientFromSpec }));
